+++ src/assets/icons/back-arrow-icon/back-arrow-icon.tsx
@@ -1,6 +1,7 @@
 import { Component, h } from '@stencil/core';
 
 @Component({
+  shadow: true,
   tag: 'mvx-back-arrow-icon',
 })
 export class BackArrowIcon {