+++ src/components/functional/ledger/ledger.types.ts
@@ -1,0 +1,44 @@
+export interface IConnectScreenData {
+  customContentMarkup?: string;
+  disabled?: boolean;
+  error?: string;
+}
+
+export interface IAccountScreenData {
+  accounts: ILedgerAccount[];
+  startIndex: number;
+  addressesPerPage: number;
+  isLoading: boolean;
+}
+
+export interface IConfirmScreenData {
+  data?: string;
+  selectedAddress: string;
+  confirmAddressText?: string;
+  authText?: string;
+  addressExplorerLink: string;
+}
+
+export interface ILedgerConnectPanelData {
+  connectScreenData: IConnectScreenData | null;
+  accountScreenData: IAccountScreenData | null;
+  shouldClose?: true;
+  confirmScreenData: IConfirmScreenData | null;
+}
+
+export interface ILedgerAccount {
+  address: string;
+  balance: string;
+  usdValue?: string;
+  index: number;
+}
+
+export enum LedgerConnectEventsEnum {
+  CONNECT_DEVICE = 'CONNECT_DEVICE',
+  ACCESS_WALLET = 'ACCESS_WALLET',
+  GO_TO_PAGE = 'GO_TO_PAGE',
+  CLOSE_LEDGER_CONNECT_PANEL = 'CLOSE_LEDGER_CONNECT_PANEL',
+  OPEN_LEDGER_CONNECT_PANEL = 'OPEN_LEDGER_CONNECT_PANEL',
+  DATA_UPDATE = 'DATA_UPDATE',
+  UI_DISCONNECTED = 'UI_DISCONNECTED',
+}